Generification courtesy of Zell, modulo my own currently still formative opinion that multi-argument generic types should not have a space. Thus:
   new HashMap<BComponent,Object>()

not

   new HashMap<BComponent, Object>().

On the other hand, I think the space is useful and necessary when declaring
generic methods and classes:

   public <T, C extends Collection<T>> void binarySearch (C c, T t)

   public class Entry<V, C extends Component>

but when actually "instantiating" a generic type, I want the type to be all one
sequence of letters, not broken into multiple space-separated blocks which
makes my brain think that it's multiple things.

Anyhow, just thinking "out loud" here.